Fix more failures
