+++ subprojects/integ-test/src/integTest/groovy/org/gradle/integtests/samples/dependencymanagement/SamplesTroubleshootingDependencyResolutionIntegrationTest.groovy
@@ -16,22 +16,17 @@
 
 package org.gradle.integtests.samples.dependencymanagement
 
-import org.gradle.integtests.fixtures.AbstractIntegrationSpec
-import org.gradle.integtests.fixtures.RepoScriptBlockUtil
+import org.gradle.integtests.fixtures.AbstractSampleIntegrationTest
 import org.gradle.integtests.fixtures.Sample
 import org.gradle.integtests.fixtures.UsesSample
 import org.junit.Rule
 
-class SamplesTroubleshootingDependencyResolutionIntegrationTest extends AbstractIntegrationSpec {
+class SamplesTroubleshootingDependencyResolutionIntegrationTest extends AbstractSampleIntegrationTest {
 
     private static final String COPY_LIBS_TASK_NAME = 'copyLibs'
 
     @Rule
     Sample sample = new Sample(testDirectoryProvider)
-
-    def setup() {
-        executer.usingInitScript(RepoScriptBlockUtil.createMirrorInitScript())
-    }
 
     @UsesSample("userguide/dependencyManagement/troubleshooting/cache/changing")
     def "can declare custom TTL for dependency with dynamic version"() {
