+++ subprojects/platform-base/src/integTest/groovy/org/gradle/language/base/LanguageTypeSampleIntegrationTest.groovy
@@ -15,7 +15,8 @@
  */
 
 package org.gradle.language.base
-import org.gradle.integtests.fixtures.AbstractSampleIntegrationTest
+
+import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.RepoScriptBlockUtil
 import org.gradle.integtests.fixtures.Sample
 import org.gradle.util.Requires
@@ -23,19 +24,18 @@ import org.gradle.util.TestPrecondition
 import org.junit.Rule
 
 @Requires(TestPrecondition.ONLINE)
-class LanguageTypeSampleIntegrationTest extends AbstractSampleIntegrationTest {
+class LanguageTypeSampleIntegrationTest extends AbstractIntegrationSpec {
     @Rule
     Sample languageTypeSample = new Sample(temporaryFolder, "customModel/languageType")
-
-    def setup() {
-        executer.usingInitScript(RepoScriptBlockUtil.createMirrorInitScript())
-    }
 
     def "shows custom language sourcesets in component"() {
         given:
         sample languageTypeSample
+        RepoScriptBlockUtil.replaceExternalRepos(testDirectory)
+
         when:
         succeeds "components"
+
         then:
         output.contains """
 DocumentationComponent 'docs'
@@ -56,10 +56,14 @@ Binaries
     def "can build binary"() {
         given:
         sample languageTypeSample
+        RepoScriptBlockUtil.replaceExternalRepos(testDirectory)
+
         when:
         succeeds "assemble"
+
         then:
         executedTasks == [":compileDocsExplodedReference", ":compileDocsExplodedUserguide", ":docsExploded", ":assemble"]
+
         and:
         languageTypeSample.dir.file("build/docs/exploded").assertHasDescendants(
                 "reference/README.txt",
